+++ CHANGELOG.md
@@ -6,7 +6,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 Please use: Added, Changed, Deprecated, Removed, Fixed, or Security as labels.
 <br>
 
-#### 0.2.2 - (Unreleased)
+#### 0.5.0 - (Unreleased)
 ##### Added
 * ...
 
@@ -25,6 +25,13 @@ Please use: Added, Changed, Deprecated, Removed, Fixed, or Security as labels.
 ##### Security
 * ...
 
+
+#### 0.4.0 - (2016-06-02)
+##### Added
+* module: format
+ * method: format.jsonToCsv
+   * converts a valid dataSet into a csv formatted string
+ * unit tests
 
 #### 0.3.0 - (2016-06-01)
 ##### Added